fix(block): Runtime db error and populate total output in block table…

+++ components/stores/blocks/src/main/resources/db/migration/h2/V0_100_1__init.sql
@@ -27,7 +27,7 @@ create table epoch
 (
     number bigint       not null primary key,
     block_count         int             null,
-    transaction_count   int             null,
+    transaction_count   bigint          null,
     total_output        bigint          null,
     start_time          bigint          null,
     end_time            bigint          null,
